Fix VAAPI build, broke upstream by https://crrev.com/c/2392140
WARNING: This is a WIP patch taken from https://crrev.com/c/2406168
which hasn't landed in ToT yet, necessary to get Brave building in
the meantime.

TODO: Replace this patch with the actual patch that lands in ToT,
or drop it when rebasing into a newer release in the dev channel.